map.js→index.jsに移植
